[#658] Seprate file for each secrets logic
Moved secrets logic classes into individual files for better
organization and maintainability.

* `CreateSecret`, `ShowSecret`, `ShowMetadata`, and `BurnSecret` are
now in their respective files.
* The main secrets file now requires these newly created files.
* Enhances readability and makes the code easier to navigate and test.
